pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools-scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "scribed"
dynamic = ["version"]
description = "Audio transcription daemon with wake word detection and voice commands"
authors = [{name = "Scribed Team", email = "team@scribed.dev"}]
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0.1",
    "watchdog>=4.0.0",
    "sounddevice>=0.4.6",
    "numpy>=1.26.0",
    "requests>=2.32.0",
    "click>=8.1.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "pre-commit>=3.4.0",
    "types-PyYAML>=6.0.0",
    "types-requests>=2.31.0",
    "httpx>=0.25.0",
]
whisper = [
    "openai-whisper>=20231117",
    "torch>=2.0.0,<2.5.0",
    "faster-whisper>=0.10.0",
]
openai = [
    "openai>=1.0.0",
]
wake_word = [
    "pvporcupine>=3.0.0",
    "pyaudio>=0.2.11",
]
# Audio file processing for format support
audio_processing = [
    "librosa>=0.10.0",
]

[project.scripts]
scribed = "scribed.cli:main"

[project.urls]
Homepage = "https://github.com/oldirty/scribed"
Documentation = "https://github.com/oldirty/scribed/blob/main/README.md"
Repository = "https://github.com/oldirty/scribed"
"Bug Tracker" = "https://github.com/oldirty/scribed/issues"
Changelog = "https://github.com/oldirty/scribed/releases"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.package-data]
scribed = ["py.typed"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/migrations/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.coverage.xml]
output = "coverage.xml"

[tool.setuptools_scm]
write_to = "src/scribed/_version.py"
version_scheme = "guess-next-dev"
local_scheme = "no-local-version"
